requirements.txt
# ETL Dependencies
pandas>=2.0.0
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0
requests>=2.31.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
GeoAlchemy2>=0.14.0

# API Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Data Processing
numpy>=1.24.0
scipy>=1.11.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
click>=8.1.0
schedule>=1.2.0

# Logging & Monitoring
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# Optional: For R stats19 package integration
# rpy2>=3.5.0
